(svn r11073) [0.5] -Backport from trunk (r11039, r11029, r11027, r11024, r11021, r11020, r11018): - Fix: Underflow that caused overflows in the performance rating [FS#1179] (r11039) - Fix [Windows]: MIDI does not stop when closing openttd [FS#1164] (r11029) - Fix: Do not unconditionally assume that a tile has a depot (r11027) - Fix: Give a more correct error when building some things on tile 0 [FS#1173] (r11024) - Fix: Do not display income/expenses when they do not belong to a "valid" tile, like the money cheat and giving money [FS#1175] (r11021) - Fix: One could not give money when (s)he had too much money [FS#1174] (r11020) - Fix: Disallow buying/selling shares in your own company or a bankrupt company [FS#1169] (r11018)

+++ main_gui.c
@@ -88,7 +88,7 @@ void HandleOnEditText(WindowEvent *e)
 #ifdef ENABLE_NETWORK
 	case 3: { /* Give money, you can only give money in excess of loan */
 		const Player *p = GetPlayer(_current_player);
-		int32 money = min(p->money64 - p->current_loan, atoi(e->we.edittext.str) / _currency->rate);
+		int32 money = min(p->money64 - p->current_loan, (int64)(atoi(e->we.edittext.str) / _currency->rate));
 
 		money = clamp(money, 0, 20000000); // Clamp between 20 million and 0
